Modify the AI to handle the large ship
The AI would try the opposite direction, but it wouldn't find a pattern
if the last - secondToLast was equal to 30. I added that math to the
function and it works fine now. I also changed the way it works for
waiting for the turn. Game has a function called wait that calls the
set timeout function. It just looks nicer.