Major overhaul of the enemy AI
So the previous code was working, but doing weird things. I went back to
the drawing board and started with some pseudocode. I created a player
object that the enemy extends. The enemy object contains the functions
necessary to make a move. I based the code in the enemy object off the
pseudocode that I wrote. I had to modify some things, but it seems to be
working so far. It correctly switches the opposite direction and uses
the possibleHits array correctly. It needs some further testing though.

Other than that, I still need to handle the game ending. I think I'd
also like to add a way for the player to see which ships have been sunk.
Maybe a button to view the players board. I also need to disable player
input immediately after they've made a move. At the moment, the player
can continously make moves until the timeout completes.