Change the AI logic
I changed the way the AI attempts to sink a ship. Before, it would get
the adjacent boxes and then try to follow a pattern. If it missed or the
move was invalid it would switch directions. What it does now is similar
but not exactly the same.

When it hits a ship it will get the adjacent boxes and begin looking for
a pattern. Once a pattern is establish it will build a separate list of
possible places to hit that will span up, down, left, or right
accordingly. It won't include spaces that aren't valid.

It will start working its way through that array of possible hits until
it misses. Once it misses, the missed function will filter the hits.
It will filter out the hits continuing in the direction it was going.
For example, if it was going up (-10) and missed at box 30, it will
filter out any numbers that are less than 30. That will leave it with
options that are all in the opposite direction.

Once the ship has been sunk, it will clear the adjacentBoxes and
possibleHits array which will reset it back to doing random moves. It
will only check for a pattern if there are no possible hits, and if
there are currently adjacent boxes to hit. In this configuration, the AI
would be currently searching for the ship, which means it should check
for a pattern. If it's not searching, then it doesn't need to find a
pattern.