Modify getAdjacentBoxes function
I realized that it was using the wrong isMoveValid function. Moves were
never being checked for being on the same row or not. I switched to
using the map function to add 10, -10, 1, and -1. I also use filter
to get rid of any undefined numbers. Those occur when a move isn't valid
and nothing is returned in the map function.