Modify the setPossibleHits function
Switched to using the map function to add the possible places to hit. I
realized that there can only be 10 possible places to hit once a pattern
has been found. The map function starts with an empty array of length 10
and then starts to add the possible places based on the index it's
currently on. The first ten are in the direction of the pattern, and the
next ten are in the opposite direction. It's then filtered to remove
any undefined elements, and any moves that aren't valid.